Refactor tests into a shared module to verify 2 and 3 behave the same.
