+++ test/python/proto3/conftest.py
@@ -1,4 +1,5 @@
 from pytest import fixture
+from test.python.shared.conftest import *
 import os
